document and clean stylingTool class
